requirements.txt
fastapi
uvicorn[standard]
requests
sentence-transformers
qdrant-client
pydantic
python-dotenv